requirements.txt
aiohttp==3.7.3
amqp==5.0.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.3
django-rest-auth==0.9.5
djangorestframework==3.12.2
idna==2.10
kombu==5.0.2
multidict==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
pycparser==2.20
pytz==2020.4
rabbitmq==0.2.0
requests==2.25.0
six==1.15.0
slackclient==2.9.3
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
